requirements.txt
python-telegram-bot
Command
requests
py-mon
colorama
watchdog
